+++ app/src/main/java/protect/card_locker/ScanActivity.java
@@ -36,11 +36,13 @@ public class ScanActivity extends AppCompatActivity {
     private DecoratedBarcodeView barcodeScannerView;
 
     private String cardId;
+    private String addGroup;
     private boolean torch = false;
 
     private void extractIntentFields(Intent intent) {
         final Bundle b = intent.getExtras();
-        cardId = b != null ? b.getString("cardId") : null;
+        cardId = b != null ? b.getString(LoyaltyCardEditActivity.BUNDLE_CARDID) : null;
+        addGroup = b != null ? b.getString(LoyaltyCardEditActivity.BUNDLE_ADDGROUP) : null;
         Log.d(TAG, "Scan activity: id=" + cardId);
     }
 
@@ -76,6 +78,9 @@ public class ScanActivity extends AppCompatActivity {
                 Bundle scanResultBundle = new Bundle();
                 scanResultBundle.putString(BarcodeSelectorActivity.BARCODE_CONTENTS, result.getText());
                 scanResultBundle.putString(BarcodeSelectorActivity.BARCODE_FORMAT, result.getBarcodeFormat().toString());
+                if (addGroup != null) {
+                    scanResultBundle.putString(LoyaltyCardEditActivity.BUNDLE_ADDGROUP, addGroup);
+                }
                 scanResult.putExtras(scanResultBundle);
                 ScanActivity.this.setResult(RESULT_OK, scanResult);
                 finish();
@@ -167,6 +172,9 @@ public class ScanActivity extends AppCompatActivity {
             Bundle manualResultBundle = new Bundle();
             manualResultBundle.putString(BarcodeSelectorActivity.BARCODE_CONTENTS, barcodeValues.content());
             manualResultBundle.putString(BarcodeSelectorActivity.BARCODE_FORMAT, barcodeValues.format());
+            if (addGroup != null) {
+                manualResultBundle.putString(LoyaltyCardEditActivity.BUNDLE_ADDGROUP, addGroup);
+            }
             manualResult.putExtras(manualResultBundle);
             ScanActivity.this.setResult(RESULT_OK, manualResult);
             finish();