+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Changes:
 
 - Make links in notes clickable
+- Pre-select group the user is currently in when creating a new card
 
 ## v2.1.0 (2021-08-01)